requirements.txt
azure-functions
pyGithub
 mysql-connector